# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  models.Jurnal:
    properties:
      _id:
        type: integer
      DOI:
        type: string
      E_ISSN:
        type: string
      P_ISSN:
        type: string
      aggregation_type:
        type: string
      akreditasi:
        type: string
      alamat_web_jurnal:
        type: string
      anggota_penelitian:
        type: string
      authors:
        type: string
      bahasa_id:
        type: string
      create_dosen_id:
        type: string
      created_at:
        type: string
      cron_semester:
        type: string
      cron_tahun:
        type: string
      dari_api_sinta:
        type: string
      deleted_at:
        type: string
      file_diterima:
        type: string
      file_hasil_uji_plagiarim:
        type: string
      file_penilaian_reviewer:
        type: string
      file_revisi:
        type: string
      file_selesai_dicetak:
        type: string
      file_submit:
        type: string
      file_sudah_revisi:
        type: string
      file_terpublikasi:
        type: string
      file_upload:
        type: string
      halaman_akhir:
        type: string
      halaman_awal:
        type: string
      id_sinta:
        type: string
      impact_factor:
        type: string
      indexer:
        items:
          type: string
        type: array
      is_produk:
        type: string
      isValid:
        type: string
      jenis_jurnal:
        type: string
      jml_penulis:
        type: string
      judul_artikel:
        type: string
      keterangan:
        type: string
      kode_akreditasi:
        type: string
      kode_jenis_jurnal:
        type: string
      kode_prodi:
        type: string
      kode_scope:
        type: string
      komentar:
        type: string
      mahasiswa_penelitian:
        type: string
      nama_dosen:
        type: string
      nama_fakultas:
        type: string
      nama_jurnal:
        type: string
      nama_jurusan:
        type: string
      nomor_jurnal:
        type: string
      penerbit:
        type: string
      periode:
        type: string
      posisi:
        type: string
      primaryKey:
        type: string
      produk_penelitian:
        type: string
      produk_penelitian_id:
        type: string
      produk_penelitian_judul:
        type: string
      produk_pengabdian:
        type: string
      produk_pengabdian_id:
        type: string
      produk_pengabdian_judul:
        type: string
      satuan:
        type: string
      scope:
        type: string
      semester:
        type: string
      semester_type:
        type: string
      sinta:
        type: string
      sitasi:
        type: string
      status_lengkap:
        type: string
      sumber_produk:
        type: string
      tahun_ajaran:
        type: string
      tahun_data:
        type: string
      tahun_publish:
        type: string
      tanggal:
        type: string
      tbName:
        type: string
      unit:
        $ref: '#/definitions/models.Unit'
      updated_at:
        type: string
      url_dokumen:
        type: string
      url_peer_review:
        type: string
      valid_ipk:
        type: string
      valid_ipk_komentar:
        type: string
      volume_jurnal:
        type: string
      volume_kegiatan:
        type: string
      waktu_pelaksanaan:
        type: string
    type: object
  models.KaryaAkhir:
    properties:
      _id:
        type: integer
      current_state:
        type: string
      judul:
        type: string
      main_stage:
        type: string
      nama_lengkap:
        type: string
      nama_pa:
        type: string
      nama_pembimbing_1:
        type: string
      nama_pembimbing_2:
        type: string
      nama_pembimbing_3:
        type: string
      nama_penguji_1:
        type: string
      nama_penguji_2:
        type: string
      nama_penguji_3:
        type: string
      nama_penguji_4:
        type: string
      nama_penguji_5:
        type: string
      nama_penguji_6:
        type: string
      nilai_akhir:
        type: string
      nim:
        type: string
      status_judul:
        type: string
      tahun_masuk:
        type: integer
      unit:
        $ref: '#/definitions/models.Unit'
    type: object
...
  models.ListResponse:
    properties:
      datas: {}
      pagination:
        $ref: '#/definitions/models.PaginationResponse'
      status:
        example: success
        type: string
    type: object
...
  models.PaginationResponse:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      pages:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  models.Unit:
    properties:
      fakultas:
        type: string
      fkt_kode:
        type: string
      jrs_kode:
        type: string
      jurusan:
        type: string
      prd_kode:
        type: string
      prodi:
        type: string
      uk_kode:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: anggadek857@gmail.com
    name: Dwi Angga
    url: https://github.com/anggananda
  description: |-
    This repository contains the **RESTful API** for the **Executive Information System (EIS) Undiksha**, built with **Golang** using the **Gin framework**.
    The API consumes data from **MongoDB** (populated via ETL from SQL in a separate repository) and uses **Redis** for caching.
    The system is integrated with **Undiksha Single Sign-On (SSO)** and implements **JWT Authentication** for secure access.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/anggananda
  title: Executive Information System RESTful API
  version: "1.0"
paths:
...
  /jurnal:
    get:
      consumes:
      - application/json
      description: Mendapatkan data jurnal berdasarkan filter dengan pagination
      parameters:
      - default: 1
        description: 'Nomor halaman (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Jumlah data per halaman (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter berdasarkan kode fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan kode jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan kode program studi
        in: query
        name: kodeProdi
        type: string
      - description: Filter berdasarkan tahun akademik
        in: query
        name: tahun
        type: string
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: string
      - description: Filter berdasarkan indexer
        in: query
        name: indexer
        type: string
      - description: Filter berdasarkan akreditasi
        in: query
        name: akreditasi
        type: string
      - description: Pencarian global
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListResponse'
            - properties:
                datas:
                  items:
                    $ref: '#/definitions/models.Jurnal'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get filtered jurnal
      tags:
      - Jurnal
  /jurnal/export-csv:
    get:
      consumes:
      - application/json
      description: Mengekspor daftar jurnal yang telah difilter ke dalam file CSV.
      parameters:
      - description: Maksimal data yang akan diekspor
        in: query
        name: limit
        type: integer
      - description: Filter berdasarkan Kode Fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan Kode Jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan Kode Program Studi
        in: query
        name: kodeProdi
        type: string
      - description: 'Filter berdasarkan Tahun Ajaran (default: tahun sekarang)'
        in: query
        name: tahun
        type: string
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: string
      - description: Filter berdasarkan indexer
        in: query
        name: indexer
        type: string
      - description: Filter berdasarkan akreditasi
        in: query
        name: akreditasi
        type: string
      - description: Pencarian bebas
        in: query
        name: search
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File CSV berhasil diunduh
          schema:
            type: file
        "500":
          description: Kesalahan pada server saat pengambilan data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export Jurnal ke CSV
      tags:
      - Jurnal
  /karya-akhir:
    get:
      consumes:
      - application/json
      description: Mendapatkan data karya akhir berdasarkan filter dengan pagination
      parameters:
      - default: 1
        description: 'Nomor halaman (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Jumlah data per halaman (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter berdasarkan kode fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan kode jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan kode program studi
        in: query
        name: kodeProdi
        type: string
      - description: Filter berdasarkan tahun akademik
        in: query
        name: tahun
        type: integer
      - description: Pencarian global
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListResponse'
            - properties:
                datas:
                  items:
                    $ref: '#/definitions/models.KaryaAkhir'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get filtered karya akhir
      tags:
      - Karya Akhir
  /karya-akhir/export-csv:
    get:
      consumes:
      - application/json
      description: Mengekspor daftar karya akhir yang telah difilter ke dalam file
        CSV.
      parameters:
      - description: Maksimal data yang akan diekspor
        in: query
        name: limit
        type: integer
      - description: Filter berdasarkan Kode Fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan Kode Jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan Kode Program Studi
        in: query
        name: kodeProdi
        type: string
      - description: 'Filter berdasarkan Tahun Ajaran (default: tahun sekarang)'
        in: query
        name: tahun
        type: string
      - description: Pencarian bebas
        in: query
        name: search
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File CSV berhasil diunduh
          schema:
            type: file
        "500":
          description: Kesalahan pada server saat pengambilan data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export Karya Akhir ke CSV
      tags:
      - Karya Akhir
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"